PEP 8 Changes. Nodeconection.run -> line.encode() remove beacause socket.recv returns bytes not str. New file structuring.
