+++ .gitignore
@@ -1,2 +1,4 @@
 *~
 *.pyc
+venv/
+.ideia/
