+++ examples/test_avans_network2.py
@@ -13,7 +13,7 @@ import time
 import sys
 import random
 
-sys.path.insert(0, '../src') # Import the files where the modules are located
+sys.path.insert(0, '../p2pnet') # Import the files where the modules are located
 
 from AvansNode import AvansNode
